src/main/java/com/example/sell/service/impl/InfoServiceImpl.java
package com.example.sell.service.impl;

import com.example.sell.bean.ProductInfo;
import com.example.sell.dao.ProductInfoRepository;
import com.example.sell.enums.ProductStatusEnum;
import com.example.sell.service.ProductInfoService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @ClassName InfoServiceImpl
 * @Author nihui
 * @Date 2019/5/11 16:38
 * @Version 1.0
 * @Description TODO
 */
@Service
public class InfoServiceImpl implements ProductInfoService {

    @Autowired
    private ProductInfoRepository productInfoRepository;
    @Override
    public ProductInfo findOne(String productId) {
        return productInfoRepository.findOne(productId);
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return productInfoRepository.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return productInfoRepository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return productInfoRepository.save(productInfo);
    }
}
